Address quant tooling review feedback
