+++ scripts/skippy-bf16-to-quant-layer-package.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-skippy_quantize_bin="${SKIPPY_QUANTIZE_BIN:-target/release/skippy-quantize}"
+repo_root="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
+skippy_quantize_bin="${SKIPPY_QUANTIZE_BIN:-$repo_root/target/release/skippy-quantize}"
 if [[ ! -x "$skippy_quantize_bin" ]]; then
   echo "missing executable: $skippy_quantize_bin" >&2
   echo "build it with: just skippy-quantize-standalone-release-build" >&2
